Change SourceForge links to Github
This updates generic download or discussion links.
It keeps direct links to discussions on SourceForge.

Some installation documents have not been updated in all translations.
Those will link to Github only. The secondary location at Github will be
linked once they are updated as part of the translation process.

This does not update the direct download links yet as this will be done as
part of #359.
This does not update the page footer yet as more involved changes are
planned as part of #345.