Start page: Add download links for 3.7.0
This replaces the SourceForge link to platform-specific direct links on
Github for better usability.
This requires text changes. To avoid delaying the release further, it
was decided to not run those texts through translators again.
This could still be updated post-release, of course.

This required CSS changes because the styling was previously specified
for a single link identified by an id=. This obviously breaks with
multiple links.

As variable expansion is not possible in front matter, the variables are
reorganized.

Fixes #359.