Update en-Contribution.md
Note: The hope is that this page looks easy and non-threatening, doesn’t talk about forking Git repos, pull requests etc. That information has now moved to the relevant README.md file for the website repo.

This is the top-level “overview” page for people following the “get involved” link on the website footer or similar. This page points people to the next stage, depending on what they want. It’s also the non-coders equivalent of README.md.