Create 2025-01-22-RaspMeasure
This project describes the possibility of using a Raspberry Pi, which is connected to a microphone and headphones, to make music online with other musicians worldwide via an Internet connection. The open source software Jamulus is used for this purpose. 

As part of this project, measurement methods were found to assess the transmission quality. In addition, a measuring device was developed with which the transmission quality can be easily measured.